refactor(text): split textBlueprints into focused modules
Layer 3 audit F10 — `text/textBlueprints.ts` (304 LOC) mixed four
concerns: font registry, glyph-to-blueprint conversion, 3D text
sketching, and font/text metrics. The file's only sibling was a README,
so the directory effectively had a single 304-LOC catch-all.

Split into four focused files:

- `fontRegistry.ts` — `loadFont`, `getFont`, FONT_REGISTER
- `textBlueprints.ts` — `textBlueprints` + glyph conversion (slim)
- `sketchText.ts` — 3D `sketchText` (depends on Sketches/CompoundSketch)
- `textMetrics.ts` — `textMetrics`, `fontMetrics`, result types

Public exports preserved via updated re-exports in `src/index.ts`,
`src/text.ts`, and the test file.
