+++ src/text.ts
@@ -7,13 +7,12 @@
  * ```
  */
 
+export { loadFont, getFont } from './text/fontRegistry.js';
+export { textBlueprints } from './text/textBlueprints.js';
+export { sketchText } from './text/sketchText.js';
 export {
-  loadFont,
-  getFont,
-  textBlueprints,
-  sketchText,
   textMetrics,
   fontMetrics,
   type TextMetricsResult,
   type FontMetricsResult,
-} from './text/textBlueprints.js';
+} from './text/textMetrics.js';
